requirements.txt
torch
torchvision
Pillow
pandas
